Week8/red_team_bias_lab/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langsmith>=0.1.0
python-dotenv>=1.0.0
